+++ react/setupTests.ts
@@ -1,0 +1,7 @@
+window.__RUNTIME__ = {
+  account: 'checkoutio',
+  workspace: 'master',
+  settings: {
+    'vtex.store': { enableOrderFormOptimization: true },
+  },
+}