+++ react/typings/global.d.ts
@@ -1,0 +1,7 @@
+import { RenderContext } from 'vtex.render-runtime'
+
+declare global {
+  interface Window {
+    __RUNTIME__: RenderContext
+  }
+}